requirements.txt
fastapi==0.115.9
uvicorn[standard]==0.34.3
python-multipart==0.0.20
python-jose[cryptography]==3.5.0
PyPDF2==3.0.1
pdfplumber==0.11.7
langchain==0.3.25
langchain-openai==0.3.23
langchain-community==0.3.25
langchain-chroma==0.2.4
openai==1.86.0
tiktoken==0.9.0
chromadb==1.0.12
libquadmath==2.3.0
pandas==2.3.0
python-dotenv==1.1.0
pydantic==2.11.7
pydantic-settings==2.9.1
requests==2.32.4
aiofiles==24.1.0
tqdm==4.67.1
loguru==0.7.3
pytest==8.4.0
pytest-asyncio==1.0.0
httpx==0.28.1
SQLAlchemy~=2.0.41